Fix (brew data validation): ensure numeric fields like dose are parsed as integers

+++ server/middleware/validateBrewData.js
@@ -14,17 +14,20 @@ const validateBrewData = [
     .optional({ values: 'falsy' })
     .isNumeric()
     .withMessage('Dose must be a number')
-    .isInt({ min: 1 }),
+    .isInt({ min: 1 })
+    .toInt(),
   body('yield', 'Yield must be bigger than 0')
     .optional({ values: 'falsy' })
     .isNumeric()
     .withMessage('Yield must be a number')
-    .isInt({ min: 1 }),
+    .isInt({ min: 1 })
+    .toInt(),
   body('temperature', 'Temperature must be between 1 and 100 (Celsius)')
     .optional({ values: 'falsy' })
     .isNumeric()
     .withMessage('Temperature must be a number')
-    .isInt({ min: 0, max: 100 }),
+    .isInt({ min: 0, max: 100 })
+    .toInt(),
   body('rating', 'Rating must be between 0 to 5').isInt({ min: 0, max: 5 }),
   body('notes').optional({ values: 'falsy' }).trim().escape(),
 ]
